fix(macos): observe fast child process exits reliably
