+++ apps/macos/Sources/OpenClaw/ShellExecutor.swift
@@ -103,55 +103,6 @@ enum ShellExecutor {
         }
     }
 
-    private final class ProcessExitSignal: @unchecked Sendable {
-        private let lock = NSLock()
-        private let source: DispatchSourceProcess
-        private var continuation: CheckedContinuation<Void, Never>?
-        private var finished = false
-
-        init(processIdentifier: pid_t) {
-            self.source = DispatchSource.makeProcessSource(
-                identifier: processIdentifier,
-                eventMask: .exit,
-                queue: .global(qos: .userInitiated))
-            self.source.setEventHandler { [weak self] in
-                self?.finish()
-            }
-            self.source.resume()
-        }
-
-        func wait() async {
-            await withTaskCancellationHandler {
-                await withCheckedContinuation { continuation in
-                    self.lock.lock()
-                    guard !self.finished else {
-                        self.lock.unlock()
-                        continuation.resume()
-                        return
-                    }
-                    self.continuation = continuation
-                    self.lock.unlock()
-                }
-            } onCancel: {
-                self.finish()
-            }
-        }
-
-        private func finish() {
-            self.lock.lock()
-            guard !self.finished else {
-                self.lock.unlock()
-                return
-            }
-            self.finished = true
-            let continuation = self.continuation
-            self.continuation = nil
-            self.lock.unlock()
-            self.source.cancel()
-            continuation?.resume()
-        }
-    }
-
     private static func environment(from values: [String: String]?) -> Environment {
         guard let values else { return .inherit }
         var converted: [Environment.Key: String] = [:]
@@ -259,8 +210,10 @@ enum ShellExecutor {
     {
         let processIdentifier = pid_t(execution.processIdentifier.value)
         return await withTaskCancellationHandler {
+            let exitSignal = ChildProcessExit(
+                processIdentifier: processIdentifier,
+                queue: .global(qos: .userInitiated))
             let deadline = await withTaskGroup(of: DeadlineOutcome.self) { group in
-                let exitSignal = ProcessExitSignal(processIdentifier: processIdentifier)
                 group.addTask {
                     await exitSignal.wait()
                     return .exited
@@ -277,7 +230,7 @@ enum ShellExecutor {
                 return await group.next() ?? .exited
             }
 
-            guard deadline == .timedOut else { return false }
+            guard deadline == .timedOut, !exitSignal.hasExited() else { return false }
             try? execution.send(signal: .terminate, toProcessGroup: true)
             try? await Task.sleep(for: .milliseconds(100))
             // The group leader may have exited on TERM. Keep the body alive until
